Update README.md
Coin Toss

Write a void function named coinToss that simulates the tossing of a coin (it does not need any parameters). When you call the function, it should generate a random number in the range of 1 through 2. If the random number is 1, the function should display “heads.” If the random number is 2, the function should display “tails.”

Demonstrate the function in a program that asks the user how many times the coin should be tossed and then simulates the tossing of the coin that number of times.

Hint: The random number generator should be seeded using srand in the main function only once.